refactor: exception 패키지 수정
