+++ src/main/java/wooteco/subway/controller/ExceptionController.java
@@ -4,11 +4,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import wooteco.subway.exceptions.LineDuplicationException;
-import wooteco.subway.exceptions.LineNotFoundException;
-import wooteco.subway.exceptions.NotAddableSectionException;
-import wooteco.subway.exceptions.StationDuplicationException;
-import wooteco.subway.exceptions.StationNotFoundException;
+import wooteco.subway.exception.LineDuplicationException;
+import wooteco.subway.exception.LineNotFoundException;
+import wooteco.subway.exception.NotAddableSectionException;
+import wooteco.subway.exception.StationDuplicationException;
+import wooteco.subway.exception.StationNotFoundException;
 
 @ControllerAdvice
 public class ExceptionController {
