+++ src/main/java/wooteco/subway/exception/LineDuplicationException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exceptions;
+package wooteco.subway.exception;
 
 public class LineDuplicationException extends RuntimeException {
