+++ src/main/java/wooteco/subway/exception/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exceptions;
+package wooteco.subway.exception;
 
 public class StationNotFoundException extends RuntimeException {
