+++ src/test/java/wooteco/subway/repository/LineDaoTest.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.exceptions.LineNotFoundException;
+import wooteco.subway.exception.LineNotFoundException;
 
 @JdbcTest
 @DisplayName("노선 레포지토리 레이어 테스트")
@@ -112,7 +112,6 @@ class LineDaoTest {
 
         // when
         lineDao.deleteById(id);
-
 
         // then
         assertThatThrownBy(() -> {
